feat(core): Add critial e2e tests file
fix(react): add missing testing dependency for libraries

update
