+++ e2e/esbuild/src/esbuild-essential.test.ts
@@ -1,0 +1,158 @@
+/**
+ * GOLDEN TEST SUITE — CRITICAL HEALTH CHECK
+ *
+ * This file contains the “essential” (aka golden) tests for the
+ * esbuild package. It tests the core public API.
+ *
+ * FAILURE POLICY
+ *  - If any test here fails, the package is considered broken.
+ *  - Fix the underlying issue immediately, do not merge or release until green.
+ *  - Do not update snapshots here without an accompanying, intentional code change.
+ *
+ * GUIDELINES
+ *  - Keep tests small and focused on fundamental behavior.
+ *  - No edge-cases we should only test the critical paths.
+ *  - CI pipelines must block on failures in this file (no skips or timeouts).
+ *
+ * MAINTENANCE
+ *  - Whenever you change core behavior, update this file first to cover the new expectations.
+ *  - Add new “essential” tests here only if they test core functionality.
+ * ──────────────────────────────────────────────────────────────────────────────
+ */
+import {
+  checkFilesDoNotExist,
+  checkFilesExist,
+  cleanupProject,
+  detectPackageManager,
+  newProject,
+  packageManagerLockFile,
+  readFile,
+  readJson,
+  runCLI,
+  runCommand,
+  tmpProjPath,
+  uniq,
+  updateFile,
+  updateJson,
+} from '@nx/e2e/utils';
+import { join } from 'path';
+describe('Esbuild essential tests', () => {
+  beforeEach(() => newProject({ packages: ['@nx/js', '@nx/esbuild'] }));
+
+  afterEach(() => cleanupProject());
+  it('should setup and build projects using build', async () => {
+    const myPkg = uniq('my-pkg');
+    runCLI(
+      `generate @nx/js:lib ${myPkg} --directory=libs/${myPkg} --bundler=esbuild`
+    );
+    updateFile(`libs/${myPkg}/src/index.ts`, `console.log('Hello');\n`);
+    updateJson(join('libs', myPkg, 'project.json'), (json) => {
+      json.targets.build.options.assets = [`libs/${myPkg}/assets/*`];
+      return json;
+    });
+    updateFile(`libs/${myPkg}/assets/a.md`, 'file a');
+    updateFile(`libs/${myPkg}/assets/b.md`, 'file b');
+
+    // Copy package.json as asset rather than generate with Nx-detected fields.
+    runCLI(`build ${myPkg} --generatePackageJson=false`);
+    const packageJson = readJson(`libs/${myPkg}/package.json`);
+    // This is the file that is generated by lib generator (no deps, no main, etc.).
+    expect(packageJson).toEqual({
+      name: `@proj/${myPkg}`,
+      version: '0.0.1',
+      private: true,
+      type: 'commonjs',
+      main: './index.cjs',
+      types: './index.d.ts',
+      dependencies: {},
+    });
+
+    // Build normally with package.json generation.
+    runCLI(`build ${myPkg}`);
+
+    expect(runCommand(`node dist/libs/${myPkg}/index.cjs`)).toMatch(/Hello/);
+    // main field should be set correctly in package.json
+    checkFilesExist(
+      `dist/libs/${myPkg}/package.json`,
+      `dist/libs/${myPkg}/${
+        packageManagerLockFile[detectPackageManager(tmpProjPath())]
+      }`
+    );
+    expect(runCommand(`node dist/libs/${myPkg}`)).toMatch(/Hello/);
+
+    expect(runCommand(`node dist/libs/${myPkg}/index.cjs`)).toMatch(/Hello/);
+    // main field should be set correctly in package.json
+
+    expect(readFile(`dist/libs/${myPkg}/assets/a.md`)).toMatch(/file a/);
+    expect(readFile(`dist/libs/${myPkg}/assets/b.md`)).toMatch(/file b/);
+
+    /* Metafile is not generated by default, but passing --metafile generates it.
+     */
+    checkFilesDoNotExist(`dist/libs/${myPkg}/meta.json`);
+    runCLI(`build ${myPkg} --metafile`);
+    checkFilesExist(`dist/libs/${myPkg}/meta.json`);
+
+    /* Type errors are turned on by default
+     */
+    updateFile(
+      `libs/${myPkg}/src/index.ts`,
+      `
+      const x: number = 'a'; // type error
+      console.log('Bye');
+    `
+    );
+    expect(() => runCLI(`build ${myPkg}`)).toThrow();
+    expect(() =>
+      runCLI(`build ${myPkg} --skipTypeCheck --no-declaration`)
+    ).not.toThrow();
+    expect(runCommand(`node dist/libs/${myPkg}/index.cjs`)).toMatch(/Bye/);
+    // Reset file
+    updateFile(
+      `libs/${myPkg}/src/index.ts`,
+      `
+      console.log('Hello');
+    `
+    );
+  }, 300_000);
+
+  it('should support non-bundle builds', () => {
+    const myPkg = uniq('my-pkg');
+    runCLI(
+      `generate @nx/js:lib ${myPkg} --directory=libs/${myPkg} --bundler=esbuild`
+    );
+    updateFile(`libs/${myPkg}/src/lib/${myPkg}.ts`, `console.log('Hello');\n`);
+    updateFile(`libs/${myPkg}/src/index.ts`, `import './lib/${myPkg}.cjs';\n`);
+
+    runCLI(`build ${myPkg} --bundle=false`);
+
+    checkFilesExist(
+      `dist/libs/${myPkg}/libs/${myPkg}/src/lib/${myPkg}.cjs`,
+      `dist/libs/${myPkg}/index.cjs`
+    );
+    // Test files are excluded in tsconfig (e.g. tsconfig.lib.json)
+    checkFilesDoNotExist(
+      `dist/libs/${myPkg}/libs/${myPkg}/src/lib/${myPkg}.spec.cjs`
+    );
+    // Can run package (package.json fields are correctly generated)
+    expect(runCommand(`node dist/libs/${myPkg}`)).toMatch(/Hello/);
+  }, 300_000);
+
+  it('should support external esbuild.config.js file', async () => {
+    const myPkg = uniq('my-pkg');
+    runCLI(
+      `generate @nx/js:lib ${myPkg} --directory=libs/${myPkg} --bundler=esbuild`
+    );
+    updateFile(
+      `libs/${myPkg}/esbuild.config.js`,
+      `console.log('custom config loaded');\nmodule.exports = {};\n`
+    );
+    updateJson(join('libs', myPkg, 'project.json'), (json) => {
+      delete json.targets.build.options.esbuildOptions;
+      json.targets.build.options.esbuildConfig = `libs/${myPkg}/esbuild.config.js`;
+      return json;
+    });
+
+    const output = runCLI(`build ${myPkg}`);
+    expect(output).toContain('custom config loaded');
+  }, 120_000);
+});
