+++ e2e/next/src/next.test.ts
@@ -33,40 +33,6 @@ describe('Next.js Applications', () => {
 
   afterAll(() => cleanupProject());
 
-  it('should support generating projects with the new name and root format', () => {
-    const appName = uniq('app1');
-    const libName = uniq('@my-org/lib1');
-
-    runCLI(
-      `generate @nx/next:app ${appName} --no-interactive --linter=eslint --unitTestRunner=jest`
-    );
-
-    // check files are generated without the layout directory ("apps/") and
-    // using the project name as the directory when no directory is provided
-    checkFilesExist(`${appName}/src/app/page.tsx`);
-    // check build works
-    expect(runCLI(`build ${appName}`)).toContain(
-      `Successfully ran target build for project ${appName}`
-    );
-    // check tests pass
-    const appTestResult = runCLI(`test ${appName} --passWithNoTests`);
-    expect(appTestResult).toContain(
-      `Successfully ran target test for project ${appName}`
-    );
-
-    runCLI(
-      `generate @nx/next:lib ${libName} --buildable --no-interactive --linter=eslint --unitTestRunner=jest`
-    );
-
-    // check files are generated without the layout directory ("libs/") and
-    // using the project name as the directory when no directory is provided
-    checkFilesExist(`${libName}/src/index.ts`);
-    // check build works
-    expect(runCLI(`build ${libName}`)).toContain(
-      `Successfully ran target build for project ${libName}`
-    );
-  }, 600_000);
-
   it('should build in dev mode without errors', async () => {
     const appName = uniq('app');
