+++ e2e/rollup/src/rollup.test.ts
@@ -15,96 +15,6 @@ describe('Rollup Plugin', () => {
   beforeAll(() => newProject({ packages: ['@nx/rollup', '@nx/js'] }));
   afterAll(() => cleanupProject());
 
-  it('should be able to setup project to build node programs with rollup and different compilers', async () => {
-    const myPkg = uniq('my-pkg');
-    runCLI(
-      `generate @nx/js:lib ${myPkg} --directory=libs/${myPkg} --bundler=rollup`
-    );
-    updateFile(`libs/${myPkg}/src/index.ts`, `console.log('Hello');\n`);
-
-    // babel (default)
-    runCLI(
-      `generate @nx/rollup:configuration ${myPkg} --tsConfig=./tsconfig.lib.json --main=./src/index.ts`
-    );
-    updateFile(
-      `libs/${myPkg}/rollup.config.cjs`,
-      `
-      const { withNx } = require('@nx/rollup/with-nx');
-      module.exports =  withNx({
-        outputPath: '../../dist/libs/${myPkg}',
-        main: './src/index.ts',
-        tsConfig: './tsconfig.lib.json',
-        compiler: 'babel',
-        generateExportsField: true,
-        additionalEntryPoints: ['./src/{foo,bar}.ts'],
-        format: ['cjs', 'esm']
-      });
-    `
-    );
-    rmDist();
-    runCLI(`build ${myPkg}`);
-    checkFilesExist(`dist/libs/${myPkg}/index.cjs.d.ts`);
-    expect(readJson(`dist/libs/${myPkg}/package.json`).exports).toEqual({
-      '.': {
-        module: './index.esm.js',
-        import: './index.cjs.mjs',
-        default: './index.cjs.js',
-        types: './index.esm.d.ts',
-      },
-      './package.json': './package.json',
-    });
-    let output = runCommand(`node dist/libs/${myPkg}/index.cjs.js`);
-    expect(output).toMatch(/Hello/);
-
-    // swc
-    runCLI(
-      `generate @nx/rollup:configuration ${myPkg} --tsConfig=./tsconfig.lib.json --main=./src/index.ts --compiler=swc`
-    );
-    updateFile(
-      `libs/${myPkg}/rollup.config.cjs`,
-      `
-      const { withNx } = require('@nx/rollup/with-nx');
-      module.exports =  withNx({
-        outputPath: '../../dist/libs/${myPkg}',
-        main: './src/index.ts',
-        tsConfig: './tsconfig.lib.json',
-        compiler: 'swc',
-        generateExportsField: true,
-        additionalEntryPoints: ['./src/{foo,bar}.ts'],
-        format: ['cjs', 'esm']
-      });
-    `
-    );
-    rmDist();
-    runCLI(`build ${myPkg}`);
-    output = runCommand(`node dist/libs/${myPkg}/index.cjs.js`);
-    expect(output).toMatch(/Hello/);
-
-    // tsc
-    runCLI(
-      `generate @nx/rollup:configuration ${myPkg} --tsConfig=./tsconfig.lib.json --main=./src/index.ts --compiler=tsc`
-    );
-    updateFile(
-      `libs/${myPkg}/rollup.config.cjs`,
-      `
-      const { withNx } = require('@nx/rollup/with-nx');
-      module.exports =  withNx({
-        outputPath: '../../dist/libs/${myPkg}',
-        main: './src/index.ts',
-        tsConfig: './tsconfig.lib.json',
-        compiler: 'tsc',
-        generateExportsField: true,
-        additionalEntryPoints: ['./src/{foo,bar}.ts'],
-        format: ['cjs', 'esm']
-      });
-    `
-    );
-    rmDist();
-    runCLI(`build ${myPkg}`);
-    output = runCommand(`node dist/libs/${myPkg}/index.cjs.js`);
-    expect(output).toMatch(/Hello/);
-  }, 500000);
-
   it('should support additional entry-points and sourcemaps', async () => {
     const myPkg = uniq('my-pkg');
     runCLI(
