+++ e2e/vue/src/vue.test.ts
@@ -18,28 +18,6 @@ describe('Vue Plugin', () => {
 
   afterAll(() => cleanupProject());
 
-  it('should serve application in dev mode', async () => {
-    const app = uniq('app');
-
-    runCLI(
-      `generate @nx/vue:app ${app} --unitTestRunner=vitest --e2eTestRunner=playwright`
-    );
-    let result = runCLI(`test ${app}`);
-    expect(result).toContain(`Successfully ran target test for project ${app}`);
-
-    result = runCLI(`build ${app}`);
-    expect(result).toContain(
-      `Successfully ran target build for project ${app}`
-    );
-
-    // TODO: enable this when tests are passing again.
-    // if (runE2ETests()) {
-    //   const e2eResults = runCLI(`e2e ${app}-e2e --no-watch`);
-    //   expect(e2eResults).toContain('Successfully ran target e2e');
-    //   expect(await killPorts()).toBeTruthy();
-    // }
-  }, 200_000);
-
   it('should serve application in dev mode with rsbuild', async () => {
     const app = uniq('app');
