+++ e2e/webpack/src/webpack.test.ts
@@ -135,86 +135,6 @@ describe('Webpack Plugin', () => {
     expect(output).toContain('Babel env is babelEnv');
   }, 500_000);
 
-  it('should be able to build with NxWebpackPlugin and a standard webpack config file', () => {
-    const appName = uniq('app');
-    runCLI(
-      `generate @nx/web:app ${appName} --bundler webpack --directory=apps/${appName}`
-    );
-    updateFile(`apps/${appName}/src/main.ts`, `console.log('Hello');\n`);
-    updateFile(`apps/${appName}/src/foo.ts`, `console.log('Foo');\n`);
-    updateFile(`apps/${appName}/src/bar.ts`, `console.log('Bar');\n`);
-
-    updateFile(
-      `apps/${appName}/webpack.config.js`,
-      `
-      const path  = require('path');
-      const { NxAppWebpackPlugin } = require('@nx/webpack/app-plugin');
-
-      module.exports = {
-        target: 'node',
-        output: {
-          path: path.join(__dirname, '../../dist/apps/${appName}'),
-          // do not remove dist, so files between builds will remain
-          clean: false,
-        },
-        plugins: [
-          new NxAppWebpackPlugin({
-            compiler: 'tsc',
-            main: './src/main.ts',
-            additionalEntryPoints: [
-              './src/foo.ts',
-              {
-                entryName: 'bar',
-                entryPath: './src/bar.ts', 
-              }
-            ],
-            tsConfig: './tsconfig.app.json',
-            outputHashing: 'none',
-            optimization: false,
-          })
-        ]
-      };`
-    );
-
-    runCLI(`build ${appName}`);
-
-    expect(runCommand(`node dist/apps/${appName}/main.js`)).toMatch(/Hello/);
-    expect(runCommand(`node dist/apps/${appName}/foo.js`)).toMatch(/Foo/);
-    expect(runCommand(`node dist/apps/${appName}/bar.js`)).toMatch(/Bar/);
-
-    // Ensure dist is not removed between builds since output.clean === false
-    createFile(`dist/apps/${appName}/extra.js`);
-    runCLI(`build ${appName} --skip-nx-cache`);
-    checkFilesExist(`dist/apps/${appName}/extra.js`);
-  }, 500_000);
-
-  it('should bundle in NX_PUBLIC_ environment variables', () => {
-    const appName = uniq('app');
-    runCLI(
-      `generate @nx/web:app ${appName} --directory=apps/${appName} --bundler webpack`
-    );
-
-    checkFilesExist(`apps/${appName}/src/main.ts`);
-    updateFile(
-      `apps/${appName}/src/main.ts`,
-      `
-      console.log(process.env['NX_PUBLIC_TEST']);
-      `
-    );
-
-    runCLI(`build ${appName}`, {
-      env: {
-        NX_PUBLIC_TEST: 'foobar',
-      },
-    });
-
-    const mainFile = listFiles(`dist/apps/${appName}`).filter((f) =>
-      f.startsWith('main.')
-    );
-    const content = readFile(`dist/apps/${appName}/${mainFile}`);
-    expect(content).toMatch(/foobar/);
-  });
-
   it('should support babel + core-js to polyfill JS features', async () => {
     const appName = uniq('app');
     runCLI(
@@ -244,139 +164,6 @@ describe('Webpack Plugin', () => {
     updateFile(`apps/${appName}/.browserslistrc`, `IE 11\n`);
     runCLI(`build ${appName}`);
     expect(readMainFile(`dist/apps/${appName}`)).not.toMatch(`await Promise`);
-  });
-
-  it('should allow options to be passed from the executor', async () => {
-    const appName = uniq('app');
-    runCLI(
-      `generate @nx/web:app ${appName} --directory=apps/${appName} --bundler webpack`
-    );
-
-    checkFilesExist(`apps/${appName}/project.json`);
-    updateJson(`apps/${appName}/project.json`, (json) => {
-      json.targets.build = {
-        executor: '@nx/webpack:webpack',
-        outputs: ['{options.outputPath}'],
-        options: {
-          generatePackageJson: true, // This should be passed to the plugin.
-          outputPath: `dist/apps/${appName}`,
-          webpackConfig: `apps/${appName}/webpack.config.js`,
-        },
-      };
-      return json;
-    });
-
-    checkFilesExist(`apps/${appName}/webpack.config.js`);
-    updateFile(
-      `apps/${appName}/webpack.config.js`,
-      `
-        const { NxAppWebpackPlugin } = require('@nx/webpack/app-plugin');
-        const { join } = require('path');
-        module.exports = {
-          output: {
-            path: join(__dirname, '../../dist/apps/demo'),
-          },
-          plugins: [
-            new NxAppWebpackPlugin({
-              // NOTE: generatePackageJson is missing here, but executor passes it.
-              target: 'web',
-              compiler: 'swc',
-              main: './src/main.ts',
-              tsConfig: './tsconfig.app.json',
-              optimization: false,
-              outputHashing: 'none',
-            }),
-          ],
-        };`
-    );
-
-    runCLI(`build ${appName}`);
-
-    fileExists(`dist/apps/${appName}/package.json`);
-  });
-
-  it('should resolve assets from executors as relative to workspace root', () => {
-    const appName = uniq('app');
-    runCLI(
-      `generate @nx/web:app ${appName} --directory=apps/${appName} --bundler webpack`
-    );
-    updateFile('shared/docs/TEST.md', 'TEST');
-    updateJson(`apps/${appName}/project.json`, (json) => {
-      json.targets.build = {
-        executor: '@nx/webpack:webpack',
-        outputs: ['{options.outputPath}'],
-        options: {
-          assets: [
-            {
-              input: 'shared/docs',
-              glob: 'TEST.md',
-              output: '.',
-            },
-          ],
-          outputPath: `dist/apps/${appName}`,
-          webpackConfig: `apps/${appName}/webpack.config.js`,
-        },
-      };
-      return json;
-    });
-
-    runCLI(`build ${appName}`);
-
-    checkFilesExist(`dist/apps/${appName}/TEST.md`);
-  });
-
-  it('it should support building libraries and apps when buildLibsFromSource is false', () => {
-    const appName = uniq('app');
-    const myPkg = uniq('my-pkg');
-
-    runCLI(
-      `generate @nx/web:application ${appName} --directory=apps/${appName}`
-    );
-
-    runCLI(
-      `generate @nx/js:lib ${myPkg} --directory=libs/${myPkg} --importPath=@${appName}/${myPkg}`
-    );
-
-    updateFile(`libs/${myPkg}/src/index.ts`, `export const foo = 'bar';\n`);
-
-    updateFile(
-      `apps/${appName}/src/main.ts`,
-      `import { foo } from '@${appName}/${myPkg}';\nconsole.log(foo);\n`
-    );
-
-    updateFile(
-      `apps/${appName}/webpack.config.js`,
-      `
-      const path  = require('path');
-      const { NxAppWebpackPlugin } = require('@nx/webpack/app-plugin');
-
-      module.exports = {
-        target: 'node',
-        output: {
-          path: path.join(__dirname, '../../dist/${appName}')
-        },
-        plugins: [
-          new NxAppWebpackPlugin({
-            compiler: 'tsc',
-            main: 'apps/${appName}/src/main.ts',
-            tsConfig: 'apps/${appName}/tsconfig.app.json',
-            outputHashing: 'none',
-            optimization: false,
-            buildLibsFromSource: false,
-          })
-        ]
-      };`
-    );
-
-    const result = runCLI(`build ${appName}`);
-
-    expect(result).toContain(
-      `Running target build for project ${appName} and 1 task it depends on`
-    );
-    expect(result).toContain(`nx run ${myPkg}:build`);
-    expect(result).toContain(
-      `Successfully ran target build for project ${appName} and 1 task it depends on`
-    );
   });
 
   it('should be able to support webpack config with nx enhanced and babel', () => {
